modified:   Rakefile modified:   lib/sinatra.rb modified:   test/app_test.rb
Added method "resource" which takes a path and a hash of actions, options and handlers.  Seems like a cleaner and DRYer way to specify re

Added a single test in app_test.rb to test functionality of this method.

Modified Rakefile line 29 changing $SAFE from 3 to 2.  At three, it threw and error...